Move builtin role selector to separate component
